deleting component subdirectory and moving into nexus_constructor/
