+++ nexus-constructor.py
@@ -8,9 +8,7 @@ from PySide2.QtGui import QIcon
 from PySide2.QtWidgets import QApplication, QMainWindow
 from PySide2 import QtCore
 
-from nexus_constructor.component.component_type import (
-    make_dictionary_of_class_definitions,
-)
+from nexus_constructor.component_type import make_dictionary_of_class_definitions
 from nexus_constructor.main_window import MainWindow
 from nexus_constructor.model.entry import Instrument, Entry
 from nexus_constructor.model.model import Model
