+++ nexus_constructor/json/load_from_json.py
@@ -7,7 +7,7 @@ from nexus_constructor.common_attrs import (
     SHAPE_GROUP_NAME,
     PIXEL_SHAPE_GROUP_NAME,
 )
-from nexus_constructor.component.component_type import COMPONENT_TYPES
+from nexus_constructor.component_type import COMPONENT_TYPES
 from nexus_constructor.json.load_from_json_utils import (
     _find_nx_class,
     _find_attribute_from_list_or_dict,
